{"schema":"decodex/commit/1","summary":"Modularize orchestrator entrypoint and run-cycle policies","authority":"XY-1115"}
