+++ apps/decodex/src/orchestrator/entrypoints.rs
@@ -1,7 +1,5 @@
 use std::io;
 
-use state::{ConnectorBackoff, ConnectorBackoffInput};
-
 use crate::runtime;
 
 const CONTROL_PLANE_TICK_CONTEXT_FAILED_WARNING: &str = "control_plane_tick_context_failed";
@@ -20,50 +18,6 @@ pub(crate) struct McpLaneSteerRequest<'a> {
 struct ControlPlaneProjectTick {
 	snapshot: Option<OperatorStatusSnapshot>,
 	project_status: Option<OperatorProjectStatus>,
-}
-
-struct ConnectorBackoffStatusParts<'a> {
-	project_id: &'a str,
-	connector: &'a str,
-	sync_phase: &'a str,
-	quota_class: &'a str,
-	reset_unix_epoch: i64,
-	reset_source: &'a str,
-	warning: &'a str,
-	next_action: &'a str,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct StatusSnapshotCacheMiss {
-	reason: String,
-	next_action: String,
-}
-
-struct StatusSnapshotHttpResponse {
-	body: Vec<u8>,
-	published_at_unix_epoch: Option<i64>,
-}
-
-impl TrackerConnectorBackoff {
-	fn to_operator_status(
-		&self,
-		project_id: &str,
-		now_unix_epoch: i64,
-	) -> OperatorConnectorBackoffStatus {
-		operator_connector_backoff_status(
-			ConnectorBackoffStatusParts {
-				project_id,
-				connector: "linear",
-				sync_phase: self.sync_phase,
-				quota_class: self.quota_class,
-				reset_unix_epoch: self.reset_unix_epoch,
-				reset_source: self.reset_source,
-				warning: self.warning,
-				next_action: self.next_action,
-			},
-			now_unix_epoch,
-		)
-	}
 }
 
 pub(crate) fn run_once(request: RunOnceRequest<'_>) -> Result<()> {
@@ -548,272 +502,6 @@ fn resolve_private_evidence_config_path(
 	resolve_config_path(request.config_path, state_store)
 }
 
-fn status_should_attempt_operator_snapshot_cache(live: bool) -> bool {
-	!live
-}
-
-fn status_snapshot_from_local_operator_cache(
-	project: &ServiceConfig,
-	limit: usize,
-) -> std::result::Result<OperatorStatusSnapshot, StatusSnapshotCacheMiss> {
-	let response = fetch_local_operator_snapshot_response()
-		.map_err(|reason| status_snapshot_cache_miss(reason, "start or restart the local Decodex operator listener, or use `decodex status --live` for a fresh direct read"))?;
-
-	status_snapshot_from_operator_cache_response(
-		project,
-		limit,
-		response,
-		OffsetDateTime::now_utc().unix_timestamp(),
-	)
-}
-
-fn status_snapshot_from_operator_cache_response(
-	project: &ServiceConfig,
-	limit: usize,
-	response: StatusSnapshotHttpResponse,
-	now_unix_epoch: i64,
-) -> std::result::Result<OperatorStatusSnapshot, StatusSnapshotCacheMiss> {
-	let published_at_unix_epoch = response.published_at_unix_epoch.ok_or_else(|| {
-		status_snapshot_cache_miss(
-			"operator snapshot response omitted X-Decodex-Snapshot-Unix-Epoch",
-			"wait for the operator listener to publish a fresh snapshot, or use `decodex status --live`",
-		)
-	})?;
-	let snapshot_age_seconds = now_unix_epoch.saturating_sub(published_at_unix_epoch);
-
-	if snapshot_age_seconds > STATUS_OPERATOR_SNAPSHOT_MAX_AGE.as_secs() as i64 {
-		return Err(status_snapshot_cache_miss(
-			format!(
-				"operator snapshot is stale: age {snapshot_age_seconds}s exceeds {}s",
-				STATUS_OPERATOR_SNAPSHOT_MAX_AGE.as_secs()
-			),
-			"wait for the next control-plane tick or use `decodex status --live`",
-		));
-	}
-
-	let snapshot = serde_json::from_slice::<OperatorStatusSnapshot>(&response.body).map_err(|error| {
-		status_snapshot_cache_miss(
-			format!("operator snapshot JSON could not be read by this CLI: {error}"),
-			"install the matching Decodex CLI/app build or use `decodex status --live`",
-		)
-	})?;
-
-	project_status_snapshot_from_operator_cache(project, snapshot, limit, snapshot_age_seconds)
-}
-
-fn fetch_local_operator_snapshot_response()
-	-> std::result::Result<StatusSnapshotHttpResponse, String>
-{
-	let address = DEFAULT_OPERATOR_LISTEN_ADDRESS
-		.parse::<SocketAddr>()
-		.map_err(|error| format!("default operator listener address is invalid: {error}"))?;
-	let mut stream = TcpStream::connect_timeout(&address, STATUS_OPERATOR_SNAPSHOT_CONNECT_TIMEOUT)
-		.map_err(|error| format!("local operator listener is unavailable at {address}: {error}"))?;
-
-	stream
-		.set_read_timeout(Some(STATUS_OPERATOR_SNAPSHOT_IO_TIMEOUT))
-		.map_err(|error| format!("failed to set operator snapshot read timeout: {error}"))?;
-	stream
-		.set_write_timeout(Some(STATUS_OPERATOR_SNAPSHOT_IO_TIMEOUT))
-		.map_err(|error| format!("failed to set operator snapshot write timeout: {error}"))?;
-	stream
-		.write_all(
-			format!(
-				"GET {OPERATOR_APP_SNAPSHOT_ENDPOINT_PATH} HTTP/1.1\r\nHost: {DEFAULT_OPERATOR_LISTEN_ADDRESS}\r\nConnection: close\r\n\r\n"
-			)
-			.as_bytes(),
-		)
-		.map_err(|error| format!("failed to request local operator snapshot: {error}"))?;
-
-	let mut response = Vec::new();
-
-	stream
-		.read_to_end(&mut response)
-		.map_err(|error| format!("failed to read local operator snapshot: {error}"))?;
-
-	parse_operator_snapshot_http_response(&response)
-}
-
-fn parse_operator_snapshot_http_response(
-	response: &[u8],
-) -> std::result::Result<StatusSnapshotHttpResponse, String> {
-	let header_end = response
-		.windows(OPERATOR_STATE_HEADER_TERMINATOR.len())
-		.position(|window| window == OPERATOR_STATE_HEADER_TERMINATOR)
-		.ok_or_else(|| String::from("local operator snapshot response omitted HTTP headers"))?;
-	let headers = std::str::from_utf8(&response[..header_end])
-		.map_err(|error| format!("local operator snapshot headers were not UTF-8: {error}"))?;
-	let Some(status_line) = headers.lines().next() else {
-		return Err(String::from("local operator snapshot response omitted HTTP status"));
-	};
-
-	if !status_line.contains(" 200 ") {
-		return Err(format!("local operator snapshot request returned `{status_line}`"));
-	}
-
-	let published_at_unix_epoch = headers.lines().find_map(|line| {
-		line.strip_prefix("X-Decodex-Snapshot-Unix-Epoch: ")
-			.and_then(|value| value.trim().parse::<i64>().ok())
-	});
-	let body = response[header_end + OPERATOR_STATE_HEADER_TERMINATOR.len()..].to_vec();
-
-	if body.is_empty() || body.as_slice() == b"{}" {
-		return Err(String::from("local operator listener has not published a status snapshot yet"));
-	}
-
-	Ok(StatusSnapshotHttpResponse { body, published_at_unix_epoch })
-}
-
-fn project_status_snapshot_from_operator_cache(
-	project: &ServiceConfig,
-	mut snapshot: OperatorStatusSnapshot,
-	limit: usize,
-	snapshot_age_seconds: i64,
-) -> std::result::Result<OperatorStatusSnapshot, StatusSnapshotCacheMiss> {
-	if snapshot.run_limit < limit {
-		return Err(status_snapshot_cache_miss(
-			format!(
-				"operator snapshot run limit {} is lower than requested status limit {limit}",
-				snapshot.run_limit
-			),
-			"rerun with a lower `--limit`, wait for a larger published snapshot, or use `decodex status --live`",
-		));
-	}
-	if snapshot.project_id == project.service_id() {
-		mark_cached_status_snapshot(&mut snapshot, limit, snapshot_age_seconds);
-
-		return Ok(snapshot);
-	}
-
-	let Some(project_status) = snapshot
-		.projects
-		.iter()
-		.find(|status| status.project_id == project.service_id())
-		.cloned()
-	else {
-		return Err(status_snapshot_cache_miss(
-			format!(
-				"operator snapshot does not include project `{}`",
-				project.service_id()
-			),
-			"check that the local listener is serving the same registered project set, or use `decodex status --live`",
-		));
-	};
-	let mut project_snapshot = empty_control_plane_snapshot(limit);
-
-	project_snapshot.project_id = project.service_id().to_owned();
-	project_snapshot.status_source = Some(String::from("operator_snapshot_cache"));
-	project_snapshot.snapshot_age_seconds = Some(snapshot_age_seconds);
-	project_snapshot.account_control = snapshot.account_control;
-	project_snapshot.accounts = snapshot.accounts;
-	project_snapshot.projects = vec![project_status];
-	project_snapshot.connector_backoffs = snapshot
-		.connector_backoffs
-		.into_iter()
-		.filter(|backoff| backoff.project_id == project.service_id())
-		.collect();
-	project_snapshot.current_lanes = snapshot
-		.current_lanes
-		.into_iter()
-		.filter(|run| run.project_id == project.service_id())
-		.collect();
-	project_snapshot.recent_runs = snapshot
-		.recent_runs
-		.into_iter()
-		.filter(|run| run.project_id == project.service_id())
-		.collect();
-	project_snapshot.history_lanes = snapshot
-		.history_lanes
-		.into_iter()
-		.filter(|lane| lane.project_id == project.service_id())
-		.collect();
-	project_snapshot.queued_candidates = snapshot
-		.queued_candidates
-		.into_iter()
-		.filter(|candidate| candidate.project_id == project.service_id())
-		.collect();
-	project_snapshot.worktrees = snapshot
-		.worktrees
-		.into_iter()
-		.filter(|worktree| worktree.project_id == project.service_id())
-		.collect();
-	project_snapshot.post_review_lanes = snapshot
-		.post_review_lanes
-		.into_iter()
-		.filter(|lane| lane.project_id == project.service_id())
-		.collect();
-	project_snapshot.warning_details = snapshot
-		.warning_details
-		.into_iter()
-		.filter(|detail| {
-			detail.project_id.as_deref().is_none()
-				|| detail.project_id.as_deref() == Some(project.service_id())
-		})
-		.collect();
-	project_snapshot.warnings = project_warnings_from_details(&project_snapshot.warning_details);
-
-	truncate_status_snapshot_to_limit(&mut project_snapshot, limit);
-	refresh_operator_project_summary(&mut project_snapshot, None);
-
-	Ok(project_snapshot)
-}
-
-fn mark_cached_status_snapshot(
-	snapshot: &mut OperatorStatusSnapshot,
-	limit: usize,
-	snapshot_age_seconds: i64,
-) {
-	snapshot.run_limit = limit;
-	snapshot.status_source = Some(String::from("operator_snapshot_cache"));
-	snapshot.snapshot_age_seconds = Some(snapshot_age_seconds);
-
-	truncate_status_snapshot_to_limit(snapshot, limit);
-	refresh_operator_project_summary(snapshot, None);
-}
-
-fn truncate_status_snapshot_to_limit(snapshot: &mut OperatorStatusSnapshot, limit: usize) {
-	snapshot.recent_runs.truncate(limit);
-	snapshot.history_lanes.truncate(limit);
-}
-
-fn project_warnings_from_details(details: &[OperatorSnapshotWarningDetail]) -> Vec<String> {
-	let mut warnings = Vec::new();
-
-	for detail in details {
-		if !warnings.iter().any(|warning| warning == &detail.warning) {
-			warnings.push(detail.warning.clone());
-		}
-	}
-
-	warnings
-}
-
-fn status_snapshot_cache_miss(
-	reason: impl Into<String>,
-	next_action: impl Into<String>,
-) -> StatusSnapshotCacheMiss {
-	StatusSnapshotCacheMiss {
-		reason: reason.into(),
-		next_action: next_action.into(),
-	}
-}
-
-fn add_status_snapshot_cache_miss_warning(
-	snapshot: &mut OperatorStatusSnapshot,
-	project: &ServiceConfig,
-	cache_miss: StatusSnapshotCacheMiss,
-) {
-	add_operator_snapshot_warning(snapshot, STATUS_OPERATOR_SNAPSHOT_WARNING);
-
-	snapshot.warning_details.push(OperatorSnapshotWarningDetail {
-		warning: String::from(STATUS_OPERATOR_SNAPSHOT_WARNING),
-		project_id: Some(project.service_id().to_owned()),
-		repo_root: Some(project.repo_root().display().to_string()),
-		reason: cache_miss.reason,
-		next_action: Some(cache_miss.next_action),
-	});
-}
-
 fn build_live_status_command_snapshot(
 	config: &ServiceConfig,
 	workflow: &WorkflowDocument,
@@ -976,201 +664,6 @@ where
 		Err(error) if error.kind() == ErrorKind::BrokenPipe => Ok(()),
 		Err(error) => Err(error.into()),
 	}
-}
-
-fn operator_connector_backoff_status(
-	parts: ConnectorBackoffStatusParts<'_>,
-	now_unix_epoch: i64,
-) -> OperatorConnectorBackoffStatus {
-	OperatorConnectorBackoffStatus {
-		project_id: parts.project_id.to_owned(),
-		connector: parts.connector.to_owned(),
-		sync_phase: parts.sync_phase.to_owned(),
-		quota_class: parts.quota_class.to_owned(),
-		reset_at: format_optional_unix_timestamp(Some(parts.reset_unix_epoch))
-			.unwrap_or_else(|| parts.reset_unix_epoch.to_string()),
-		reset_unix_epoch: parts.reset_unix_epoch,
-		reset_source: parts.reset_source.to_owned(),
-		retry_after_seconds: parts.reset_unix_epoch.saturating_sub(now_unix_epoch).max(0),
-		next_action: parts.next_action.to_owned(),
-		warning: parts.warning.to_owned(),
-	}
-}
-
-fn connector_backoff_record_to_operator_status(
-	backoff: &ConnectorBackoff,
-	now_unix_epoch: i64,
-) -> OperatorConnectorBackoffStatus {
-	operator_connector_backoff_status(
-		ConnectorBackoffStatusParts {
-			project_id: backoff.project_id(),
-			connector: backoff.connector(),
-			sync_phase: backoff.sync_phase(),
-			quota_class: backoff.quota_class(),
-			reset_unix_epoch: backoff.reset_unix_epoch(),
-			reset_source: backoff.reset_source(),
-			warning: backoff.warning(),
-			next_action: connector_backoff_next_action(backoff.warning()),
-		},
-		now_unix_epoch,
-	)
-}
-
-fn connector_backoff_next_action(warning: &str) -> &'static str {
-	match warning {
-		TRACKER_TRANSIENT_TIMEOUT_WARNING =>
-			"Wait for the transient tracker timeout backoff; Decodex will retry tracker reads without changing lane ownership.",
-		_ => "Wait for the reset window; keep monitoring local running lanes.",
-	}
-}
-
-fn tracker_backoff_warning_label(warning: &str) -> Option<&'static str> {
-	match warning {
-		TRACKER_RATE_LIMIT_WARNING => Some(TRACKER_RATE_LIMIT_WARNING),
-		TRACKER_TRANSIENT_TIMEOUT_WARNING => Some(TRACKER_TRANSIENT_TIMEOUT_WARNING),
-		_ => None,
-	}
-}
-
-fn warnings_include_tracker_backoff(warnings: &[&str]) -> bool {
-	warnings
-		.iter()
-		.any(|warning| tracker_backoff_warning_label(warning).is_some())
-}
-
-fn snapshot_warnings_include_tracker_backoff(snapshot: &OperatorStatusSnapshot) -> bool {
-	snapshot
-		.warnings
-		.iter()
-		.any(|warning| tracker_backoff_warning_label(warning).is_some())
-}
-
-fn push_connector_backoff_warning(
-	snapshot_warnings: &mut Vec<&'static str>,
-	backoff: &OperatorConnectorBackoffStatus,
-) {
-	if let Some(warning) = tracker_backoff_warning_label(&backoff.warning)
-		&& !snapshot_warnings.contains(&warning) {
-			snapshot_warnings.push(warning);
-		}
-}
-
-fn active_stored_tracker_backoff_status(
-	state_store: &StateStore,
-	project_id: &str,
-) -> Result<Option<OperatorConnectorBackoffStatus>> {
-	let Some(backoff) = state_store.connector_backoff(project_id, "linear")? else {
-		return Ok(None);
-	};
-	let now_unix_epoch = OffsetDateTime::now_utc().unix_timestamp();
-
-	if backoff.reset_unix_epoch() <= now_unix_epoch {
-		state_store.clear_connector_backoff(project_id, "linear")?;
-
-		return Ok(None);
-	}
-
-	Ok(Some(connector_backoff_record_to_operator_status(&backoff, now_unix_epoch)))
-}
-
-fn active_stored_tracker_backoff_status_best_effort(
-	state_store: &StateStore,
-	project_id: &str,
-) -> Option<OperatorConnectorBackoffStatus> {
-	match active_stored_tracker_backoff_status(state_store, project_id) {
-		Ok(status) => status,
-		Err(error) => {
-			let _ = error;
-
-			tracing::warn!(
-				project_id = project_id,
-				"Failed to read persisted tracker backoff; sensitive runtime details were withheld."
-			);
-
-			None
-		},
-	}
-}
-
-fn persist_tracker_backoff_state(
-	state_store: &StateStore,
-	project_id: &str,
-	backoff: &TrackerConnectorBackoff,
-) {
-	if let Err(error) = state_store.upsert_connector_backoff(ConnectorBackoffInput {
-		project_id,
-		connector: "linear",
-		sync_phase: backoff.sync_phase,
-		quota_class: backoff.quota_class,
-		reset_unix_epoch: backoff.reset_unix_epoch,
-		reset_source: backoff.reset_source,
-		warning: backoff.warning,
-	}) {
-		let _ = error;
-
-		tracing::warn!(
-			project_id = project_id,
-			"Failed to persist tracker backoff; sensitive runtime details were withheld."
-		);
-	}
-}
-
-fn clear_tracker_backoff_state_best_effort(state_store: &StateStore, project_id: &str) {
-	if let Err(error) = state_store.clear_connector_backoff(project_id, "linear") {
-		let _ = error;
-
-		tracing::warn!(
-			project_id = project_id,
-			"Failed to clear persisted tracker backoff; sensitive runtime details were withheld."
-		);
-	}
-}
-
-fn render_tracker_backoff_cli_message(
-	command: &str,
-	status: &OperatorConnectorBackoffStatus,
-) -> String {
-	format!(
-		"Linear connector is in backoff for project `{}`; `{}` skipped tracker reads for `{}` until {} (retry_after_seconds={}).\n",
-		status.project_id,
-		command,
-		status.sync_phase,
-		status.reset_at,
-		status.retry_after_seconds
-	)
-}
-
-fn build_operator_status_snapshot_for_tracker_backoff(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	limit: usize,
-	status: &OperatorConnectorBackoffStatus,
-) -> Result<OperatorStatusSnapshot> {
-	let review_state_inspector = GhPullRequestReviewStateInspector {
-		github_token_env_var: Some(project.github().token_env_var().to_owned()),
-		github_command_path: project.github().command_path().map(Path::to_path_buf),
-	};
-	let mut snapshot = build_operator_status_snapshot_with_account_mode(
-		project,
-		state_store,
-		limit,
-		AccountActivityMode::Snapshot,
-	)?;
-
-	hydrate_history_lanes_from_local_ledger(project, state_store, &mut snapshot)?;
-
-	snapshot.post_review_lanes =
-		build_degraded_post_review_lane_statuses(project, state_store, &review_state_inspector)?;
-
-	add_operator_snapshot_warning(&mut snapshot, &status.warning);
-
-	snapshot.connector_backoffs.push(status.clone());
-
-	add_operator_snapshot_warning(&mut snapshot, "external_observer_status_skipped");
-	apply_terminal_history_ledger_outcomes(&mut snapshot);
-	refresh_operator_project_summary(&mut snapshot, None);
-
-	Ok(snapshot)
 }
 
 fn publish_operator_snapshot(
@@ -1724,88 +1217,6 @@ fn remember_tracker_backoff(
 	runtime.tracker_backoff = Some(backoff);
 
 	Some(status)
-}
-
-fn tracker_connector_backoff(
-	error: &Report,
-	now: Instant,
-	sync_phase: &'static str,
-) -> Option<TrackerConnectorBackoff> {
-	let message = format!("{error:#}");
-
-	if message.contains("Linear connector is rate limited") {
-		return tracker_connector_backoff_from_message(&message, now, sync_phase);
-	}
-	if message.contains("Linear connector timed out") {
-		return tracker_timeout_backoff(now, sync_phase);
-	}
-
-	None
-}
-
-fn tracker_connector_backoff_from_message(
-	message: &str,
-	now: Instant,
-	sync_phase: &'static str,
-) -> Option<TrackerConnectorBackoff> {
-	let now_unix_epoch = OffsetDateTime::now_utc().unix_timestamp();
-	let fallback_reset_unix_epoch =
-		now_unix_epoch.saturating_add(TRACKER_RATE_LIMIT_BACKOFF_SECS as i64);
-	let (reset_unix_epoch, reset_source) =
-		match parse_linear_rate_limit_reset_unix_epoch(message) {
-			Some(reset_unix_epoch) if reset_unix_epoch > now_unix_epoch =>
-				(reset_unix_epoch, "linear"),
-			_ => (fallback_reset_unix_epoch, "local_default"),
-		};
-	let retry_after_seconds = reset_unix_epoch - now_unix_epoch;
-	let retry_after_seconds = u64::try_from(retry_after_seconds).ok()?;
-
-	Some(TrackerConnectorBackoff {
-		until: now + Duration::from_secs(retry_after_seconds),
-		quota_class: "linear_graphql_rate_limit",
-		reset_unix_epoch,
-		reset_source,
-		sync_phase,
-		warning: TRACKER_RATE_LIMIT_WARNING,
-		next_action: "Wait for the reset window; keep monitoring local running lanes.",
-	})
-}
-
-fn tracker_timeout_backoff(
-	now: Instant,
-	sync_phase: &'static str,
-) -> Option<TrackerConnectorBackoff> {
-	let now_unix_epoch = OffsetDateTime::now_utc().unix_timestamp();
-	let reset_unix_epoch =
-		now_unix_epoch.saturating_add(TRACKER_TRANSIENT_TIMEOUT_BACKOFF_SECS as i64);
-
-	Some(TrackerConnectorBackoff {
-		until: now + Duration::from_secs(TRACKER_TRANSIENT_TIMEOUT_BACKOFF_SECS),
-		quota_class: "linear_graphql_timeout",
-		reset_unix_epoch,
-		reset_source: "local_transient_timeout",
-		sync_phase,
-		warning: TRACKER_TRANSIENT_TIMEOUT_WARNING,
-		next_action: "Wait for the transient tracker timeout backoff; Decodex will retry tracker reads without changing lane ownership.",
-	})
-}
-
-fn active_connector_backoff_statuses(
-	project_id: &str,
-	runtime: &ProjectDaemonRuntime,
-) -> Vec<OperatorConnectorBackoffStatus> {
-	let Some(backoff) = runtime.tracker_backoff.as_ref() else {
-		return Vec::new();
-	};
-	let now_unix_epoch = OffsetDateTime::now_utc().unix_timestamp();
-
-	vec![backoff.to_operator_status(project_id, now_unix_epoch)]
-}
-
-fn parse_linear_rate_limit_reset_unix_epoch(message: &str) -> Option<i64> {
-	let reset = message.split("rate limited until `").nth(1)?.split('`').next()?;
-
-	reset.parse().ok()
 }
 
 fn control_plane_project_deferred_snapshot(
