+++ apps/decodex/src/orchestrator/run_cycle.rs
@@ -1,16 +1,8 @@
-use state::{IssueLease, PreacquiredLeaseGuards, WORKTREE_PROVENANCE_RUNTIME_RECORDED};
+use state::{PreacquiredLeaseGuards, WORKTREE_PROVENANCE_RUNTIME_RECORDED};
 
 use crate::commit_message;
 
 const INTERNAL_RETAINED_DRAIN_MAX_PASSES: usize = 2;
-
-struct ProjectStateReconciliationContext<'a, T> {
-	tracker: &'a T,
-	project: &'a ServiceConfig,
-	workflow: &'a WorkflowDocument,
-	state_store: &'a StateStore,
-	worktree_manager: &'a WorktreeManager,
-}
 
 fn run_configured_cycle(request: RunCycleRequest<'_>) -> Result<Option<RunSummary>> {
 	let config = ServiceConfig::from_path(request.config_path)?;
@@ -276,7 +268,6 @@ fn select_recovered_retry_issue_candidate(
 
 	Ok(None)
 }
-
 fn queued_issues_for_dispatch<T>(
 	tracker: &T,
 	project: &ServiceConfig,
@@ -355,290 +346,6 @@ where
 		dry_run,
 		&next_excluded_issue_ids,
 	)
-}
-
-fn select_post_review_issue_candidate<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	excluded_issue_ids: &[&str],
-) -> Result<Option<SelectedIssueRunCandidate>>
-where
-	T: IssueTracker,
-{
-	let review_state_inspector = GhPullRequestReviewStateInspector {
-		github_token_env_var: Some(project.github().token_env_var().to_owned()),
-		github_command_path: project.github().command_path().map(Path::to_path_buf),
-	};
-
-	select_post_review_issue_candidate_with_inspector(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		excluded_issue_ids,
-		&review_state_inspector,
-	)
-}
-
-fn select_post_review_issue_candidate_with_inspector<T, I>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	excluded_issue_ids: &[&str],
-	review_state_inspector: &I,
-) -> Result<Option<SelectedIssueRunCandidate>>
-where
-	T: IssueTracker,
-	I: PullRequestReviewStateInspector,
-{
-	if let Some(issue) = select_post_review_repair_issue_candidate_with_inspector(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		excluded_issue_ids,
-		review_state_inspector,
-	)? {
-		return Ok(Some(SelectedIssueRunCandidate::new(issue, IssueDispatchMode::ReviewRepair)));
-	}
-
-	select_post_review_closeout_issue_candidate_with_inspector(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		excluded_issue_ids,
-		review_state_inspector,
-	)
-}
-
-fn select_post_review_repair_issue_candidate_with_inspector<T, I>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	excluded_issue_ids: &[&str],
-	review_state_inspector: &I,
-) -> Result<Option<TrackerIssue>>
-where
-	T: IssueTracker,
-	I: PullRequestReviewStateInspector,
-{
-	let lanes = build_post_review_lane_statuses(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		review_state_inspector,
-	)?;
-	let candidate_issue_ids = lanes
-		.iter()
-		.filter(|lane| lane.classification == "needs_review_repair")
-		.filter(|lane| !excluded_issue_ids.contains(&lane.issue_id.as_str()))
-		.map(|lane| lane.issue_id.clone())
-		.collect::<Vec<_>>();
-
-	if candidate_issue_ids.is_empty() {
-		return Ok(None);
-	}
-
-	let issues = tracker.refresh_issues(&candidate_issue_ids)?;
-	let mut issues_by_id =
-		issues.into_iter().map(|issue| (issue.id.clone(), issue)).collect::<HashMap<_, _>>();
-
-	for lane in lanes {
-		if lane.classification != "needs_review_repair" {
-			continue;
-		}
-		if excluded_issue_ids.contains(&lane.issue_id.as_str()) {
-			continue;
-		}
-		if state_store.issue_has_active_shared_claim(project.service_id(), &lane.issue_id)? {
-			continue;
-		}
-
-		if let Some(issue) = issues_by_id.remove(&lane.issue_id) {
-			return Ok(Some(issue));
-		}
-	}
-
-	Ok(None)
-}
-
-fn select_post_review_closeout_issue_candidate_with_inspector<T, I>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	excluded_issue_ids: &[&str],
-	review_state_inspector: &I,
-) -> Result<Option<SelectedIssueRunCandidate>>
-where
-	T: IssueTracker,
-	I: PullRequestReviewStateInspector,
-{
-	let completed_state = workflow.frontmatter().tracker().resolved_completed_state();
-	let lanes = build_post_review_lane_statuses(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		review_state_inspector,
-	)?;
-	let candidate_issue_ids = lanes
-		.iter()
-		.filter(|lane| post_review_lane_is_closeout_candidate(lane, completed_state))
-		.filter(|lane| !excluded_issue_ids.contains(&lane.issue_id.as_str()))
-		.map(|lane| lane.issue_id.clone())
-		.collect::<Vec<_>>();
-
-	if candidate_issue_ids.is_empty() {
-		return Ok(None);
-	}
-
-	let issues = tracker.refresh_issues(&candidate_issue_ids)?;
-	let mut issues_by_id =
-		issues.into_iter().map(|issue| (issue.id.clone(), issue)).collect::<HashMap<_, _>>();
-
-	for lane in lanes {
-		let is_closeout_candidate = post_review_lane_is_closeout_candidate(&lane, completed_state);
-
-		if !is_closeout_candidate {
-			continue;
-		}
-		if excluded_issue_ids.contains(&lane.issue_id.as_str()) {
-			continue;
-		}
-
-		if let Some(issue) = issues_by_id.remove(&lane.issue_id) {
-			if closeout_lane_active_claim_blocks_dispatch(project, state_store, &issue)? {
-				continue;
-			}
-
-			let preferred_run_identity = retained_closeout_preferred_run_identity(
-				state_store,
-				project.service_id(),
-				&issue,
-			)?;
-
-			return Ok(Some(SelectedIssueRunCandidate {
-				issue,
-				dispatch_mode: IssueDispatchMode::Closeout,
-				preferred_run_identity,
-			}));
-		}
-	}
-
-	Ok(None)
-}
-
-fn retained_closeout_preferred_run_identity(
-	state_store: &StateStore,
-	project_id: &str,
-	issue: &TrackerIssue,
-) -> Result<Option<RetainedReviewRunIdentity>> {
-	let Some(worktree) = state_store.worktree_for_issue(&issue.id)? else {
-		return Ok(None);
-	};
-	let Some(review_handoff) =
-		state_store.review_handoff_marker(project_id, &issue.id, worktree.branch_name())?
-	else {
-		return Ok(None);
-	};
-	let identity = RetainedReviewRunIdentity {
-		run_id: review_handoff.run_id().to_owned(),
-		attempt_number: review_handoff.attempt_number(),
-	};
-
-	if retained_closeout_run_identity_is_reusable(state_store, &issue.id, &identity)?
-		|| retained_closeout_handoff_identity_is_reusable_after_parent_reconciliation(
-			state_store,
-			&issue.id,
-			&identity,
-			&worktree,
-		)? {
-		return Ok(Some(identity));
-	}
-
-	Ok(None)
-}
-
-fn retained_closeout_run_identity_is_reusable(
-	state_store: &StateStore,
-	issue_id: &str,
-	identity: &RetainedReviewRunIdentity,
-) -> Result<bool> {
-	if state_store.issue_has_retry_budget_attempt_after(issue_id, identity.attempt_number)? {
-		return Ok(false);
-	}
-
-	let Some(existing_attempt) = state_store.run_attempt(&identity.run_id)? else {
-		return Ok(true);
-	};
-
-	if existing_attempt.issue_id() != issue_id
-		|| existing_attempt.attempt_number() != identity.attempt_number
-	{
-		return Ok(false);
-	}
-
-	Ok(!matches!(existing_attempt.status(), "failed" | "interrupted" | TERMINAL_GUARDED_RUN_STATUS))
-}
-
-fn retained_closeout_handoff_identity_is_reusable_after_parent_reconciliation(
-	state_store: &StateStore,
-	issue_id: &str,
-	identity: &RetainedReviewRunIdentity,
-	worktree: &WorktreeMapping,
-) -> Result<bool> {
-	if state_store.issue_has_retry_budget_attempt_after(issue_id, identity.attempt_number)? {
-		return Ok(false);
-	}
-
-	let Some(existing_attempt) = state_store.run_attempt(&identity.run_id)? else {
-		return Ok(false);
-	};
-
-	if existing_attempt.issue_id() != issue_id
-		|| existing_attempt.attempt_number() != identity.attempt_number
-	{
-		return Ok(false);
-	}
-	if !matches!(existing_attempt.status(), "failed" | "interrupted") {
-		return Ok(false);
-	}
-	if worktree_has_retry_schedule_for_run(worktree.worktree_path(), identity)? {
-		return Ok(false);
-	}
-
-	Ok(true)
-}
-
-fn worktree_has_retry_schedule_for_run(
-	worktree_path: &Path,
-	identity: &RetainedReviewRunIdentity,
-) -> Result<bool> {
-	let Some(marker) = state::read_run_activity_marker_snapshot(worktree_path)? else {
-		return Ok(false);
-	};
-
-	Ok(marker.run_id() == identity.run_id
-		&& marker.attempt_number() == identity.attempt_number
-		&& marker.retry_kind().is_some())
-}
-
-fn post_review_lane_is_closeout_candidate(
-	lane: &OperatorPostReviewLaneStatus,
-	_completed_state: &str,
-) -> bool {
-	lane.classification == "continue" && lane.reason == "pull_request_merged_closeout_pending"
-}
-
-fn post_review_lane_is_repair_candidate(lane: &OperatorPostReviewLaneStatus) -> bool {
-	lane.classification == "needs_review_repair"
 }
 
 fn run_target_issue_once<T>(context: TargetIssueRunContext<'_, T>) -> Result<Option<RunSummary>>
@@ -1029,61 +736,6 @@ where
 	}))
 }
 
-fn select_target_post_review_repair_issue_candidate_with_inspector<T, I>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	target_issue_id: &str,
-	target_issue_reference: &str,
-	review_state_inspector: &I,
-) -> Result<Option<TrackerIssue>>
-where
-	T: IssueTracker,
-	I: PullRequestReviewStateInspector,
-{
-	let lanes = build_post_review_lane_statuses(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		review_state_inspector,
-	)?;
-	let repair_lanes =
-		lanes.into_iter().filter(post_review_lane_is_repair_candidate).collect::<Vec<_>>();
-
-	if repair_lanes.is_empty() {
-		return Ok(None);
-	}
-
-	let Some(target_lane) = repair_lanes.iter().find(|lane| lane.issue_id == target_issue_id)
-	else {
-		let visible_lanes = repair_lanes
-			.iter()
-			.map(|lane| lane.issue_identifier.as_str())
-			.collect::<Vec<_>>()
-			.join(", ");
-
-		eyre::bail!(
-			"targeted retained review repair mismatch: requested issue `{}` does not match status-visible retained review repair lane(s) `{}`",
-			target_issue_reference,
-			visible_lanes,
-		);
-	};
-	let issue_ids = [target_lane.issue_id.clone()];
-	let mut issues = tracker.refresh_issues(&issue_ids)?;
-	let Some(issue_index) = issues.iter().position(|issue| issue.id == target_lane.issue_id) else {
-		return Ok(None);
-	};
-	let issue = issues.swap_remove(issue_index);
-
-	if state_store.issue_has_active_shared_claim(project.service_id(), &issue.id)? {
-		return Ok(None);
-	}
-
-	Ok(Some(issue))
-}
-
 fn run_target_status_visible_closeout_once<T>(
 	context: TargetIssueRunContext<'_, T>,
 ) -> Result<Option<RunSummary>>
@@ -1130,71 +782,6 @@ where
 		preferred_run_identity,
 		preferred_retry_budget_base: context.preferred_retry_budget_base,
 	})
-}
-
-fn select_target_post_review_closeout_issue_candidate_with_inspector<T, I>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	target_issue_id: &str,
-	target_issue_reference: &str,
-	review_state_inspector: &I,
-) -> Result<Option<SelectedIssueRunCandidate>>
-where
-	T: IssueTracker,
-	I: PullRequestReviewStateInspector,
-{
-	let completed_state = workflow.frontmatter().tracker().resolved_completed_state();
-	let lanes = build_post_review_lane_statuses(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		review_state_inspector,
-	)?;
-	let closeout_lanes = lanes
-		.into_iter()
-		.filter(|lane| post_review_lane_is_closeout_candidate(lane, completed_state))
-		.collect::<Vec<_>>();
-
-	if closeout_lanes.is_empty() {
-		return Ok(None);
-	}
-
-	let Some(target_lane) = closeout_lanes.iter().find(|lane| lane.issue_id == target_issue_id)
-	else {
-		let visible_lanes = closeout_lanes
-			.iter()
-			.map(|lane| lane.issue_identifier.as_str())
-			.collect::<Vec<_>>()
-			.join(", ");
-
-		eyre::bail!(
-			"targeted retained closeout mismatch: requested issue `{}` does not match status-visible retained closeout lane(s) `{}`",
-			target_issue_reference,
-			visible_lanes,
-		);
-	};
-	let issue_ids = [target_lane.issue_id.clone()];
-	let mut issues = tracker.refresh_issues(&issue_ids)?;
-	let Some(issue_index) = issues.iter().position(|issue| issue.id == target_lane.issue_id) else {
-		return Ok(None);
-	};
-	let issue = issues.swap_remove(issue_index);
-
-	if closeout_lane_active_claim_blocks_dispatch(project, state_store, &issue)? {
-		return Ok(None);
-	}
-
-	let preferred_run_identity =
-		retained_closeout_preferred_run_identity(state_store, project.service_id(), &issue)?;
-
-	Ok(Some(SelectedIssueRunCandidate {
-		issue,
-		dispatch_mode: IssueDispatchMode::Closeout,
-		preferred_run_identity,
-	}))
 }
 
 fn target_issue_reuses_existing_closeout_claim<T>(
@@ -1726,552 +1313,4 @@ where
 	}
 
 	Ok(None)
-}
-
-fn reconcile_project_state<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	worktree_manager: &WorktreeManager,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	let leases = state_store.list_leases(project.service_id())?;
-	let mut worktrees = state_store.list_worktrees(project.service_id())?;
-
-	if leases.is_empty() && worktrees.is_empty() {
-		return Ok(());
-	}
-
-	clear_stale_terminal_local_worktree_mappings(project, state_store, &leases, &mut worktrees)?;
-
-	if leases.is_empty() && worktrees.is_empty() {
-		return Ok(());
-	}
-
-	let mut issue_ids = HashSet::new();
-
-	for lease in &leases {
-		issue_ids.insert(lease.issue_id().to_owned());
-	}
-	for mapping in &worktrees {
-		issue_ids.insert(mapping.issue_id().to_owned());
-	}
-
-	let refreshed_issues = tracker.refresh_issues(&issue_ids.into_iter().collect::<Vec<_>>())?;
-	let issues_by_id = refreshed_issues
-		.into_iter()
-		.map(|issue| (issue.id.clone(), issue))
-		.collect::<HashMap<_, _>>();
-	let reconciliation_context = ProjectStateReconciliationContext {
-		tracker,
-		project,
-		workflow,
-		state_store,
-		worktree_manager,
-	};
-	let now_unix_epoch = OffsetDateTime::now_utc().unix_timestamp();
-	let mut cleared_terminal_lane_issue_ids = HashSet::new();
-
-	reconcile_active_project_leases(
-		&reconciliation_context,
-		&leases,
-		&issues_by_id,
-		now_unix_epoch,
-		&mut cleared_terminal_lane_issue_ids,
-	)?;
-	cleanup_missing_orphaned_project_worktree_mappings(
-		&reconciliation_context,
-		&leases,
-		&worktrees,
-		&issues_by_id,
-	)?;
-	reconcile_orphaned_active_worktree_runs(
-		&reconciliation_context,
-		&leases,
-		&worktrees,
-		&issues_by_id,
-		now_unix_epoch,
-	)?;
-	cleanup_terminal_project_worktrees(
-		&reconciliation_context,
-		&worktrees,
-		&issues_by_id,
-		&mut cleared_terminal_lane_issue_ids,
-	)?;
-
-	Ok(())
-}
-
-fn clear_stale_terminal_local_worktree_mappings(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	leases: &[IssueLease],
-	worktrees: &mut Vec<WorktreeMapping>,
-) -> Result<()> {
-	let active_issue_ids =
-		leases.iter().map(|lease| lease.issue_id().to_owned()).collect::<HashSet<_>>();
-	let mut cleared_issue_ids = Vec::new();
-
-	for mapping in worktrees.iter() {
-		if !worktree_mapping_is_stale_terminal_local_residue(
-			project,
-			state_store,
-			mapping,
-			&active_issue_ids,
-		)? {
-			continue;
-		}
-
-		state_store.clear_worktree(mapping.issue_id())?;
-
-		tracing::info!(
-			project_id = project.service_id(),
-			issue_id = mapping.issue_id(),
-			provenance_source = mapping.provenance().source(),
-			"Cleared stale terminal local worktree mapping before tracker refresh."
-		);
-
-		cleared_issue_ids.push(mapping.issue_id().to_owned());
-	}
-
-	if !cleared_issue_ids.is_empty() {
-		worktrees.retain(|mapping| {
-			!cleared_issue_ids.iter().any(|issue_id| issue_id == mapping.issue_id())
-		});
-	}
-
-	Ok(())
-}
-
-fn looks_like_tracker_issue_identifier_key(value: &str) -> bool {
-	let Some((prefix, number)) = value.rsplit_once('-') else {
-		return false;
-	};
-
-	!prefix.is_empty()
-		&& !number.is_empty()
-		&& prefix
-			.chars()
-			.all(|character| character.is_ascii_uppercase() || character.is_ascii_digit())
-		&& number.chars().all(|character| character.is_ascii_digit())
-}
-
-fn local_run_attempt_status_is_terminal(status: &str) -> bool {
-	matches!(
-		status,
-		"succeeded" | "failed" | "interrupted" | "terminated" | TERMINAL_GUARDED_RUN_STATUS
-	)
-}
-
-fn reconcile_active_project_leases<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	leases: &[IssueLease],
-	issues_by_id: &HashMap<String, TrackerIssue>,
-	now_unix_epoch: i64,
-	cleared_terminal_lane_issue_ids: &mut HashSet<String>,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	for lease in leases {
-		if reconcile_success_retained_review_lease(context, lease, issues_by_id)? {
-			continue;
-		}
-		if reconcile_terminal_retained_closeout_lease(
-			context,
-			lease,
-			issues_by_id,
-			now_unix_epoch,
-			cleared_terminal_lane_issue_ids,
-		)? {
-			continue;
-		}
-
-		reconcile_stale_project_lease(
-			context,
-			lease,
-			issues_by_id,
-			cleared_terminal_lane_issue_ids,
-		)?;
-	}
-
-	Ok(())
-}
-
-fn reconcile_success_retained_review_lease<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	lease: &IssueLease,
-	issues_by_id: &HashMap<String, TrackerIssue>,
-) -> Result<bool>
-where
-	T: IssueTracker,
-{
-	if let Some(issue) = issues_by_id.get(lease.issue_id())
-		&& issue.state.name == context.workflow.frontmatter().tracker().success_state()
-		&& retained_review_lease_matches_run(context.state_store, lease)?
-	{
-		mark_run_attempt_if_active(context.state_store, lease.run_id(), "succeeded")?;
-
-		context.state_store.clear_lease(lease.issue_id())?;
-
-		return Ok(true);
-	}
-
-	Ok(false)
-}
-
-fn reconcile_terminal_retained_closeout_lease<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	lease: &IssueLease,
-	issues_by_id: &HashMap<String, TrackerIssue>,
-	now_unix_epoch: i64,
-	cleared_terminal_lane_issue_ids: &mut HashSet<String>,
-) -> Result<bool>
-where
-	T: IssueTracker,
-{
-	let Some(issue) = issues_by_id.get(lease.issue_id()) else {
-		return Ok(false);
-	};
-
-	if !terminal_issue_keeps_retained_closeout(
-		context.tracker,
-		issue,
-		context.project,
-		context.workflow,
-		context.state_store,
-	)? {
-		return Ok(false);
-	}
-	if retained_closeout_lease_has_fresh_activity(lease, issue, context.project, now_unix_epoch)? {
-		return Ok(true);
-	}
-
-	clear_terminal_lane_labels_once(
-		context.tracker,
-		context.project,
-		issue,
-		cleared_terminal_lane_issue_ids,
-	)?;
-	mark_run_attempt_if_active(context.state_store, lease.run_id(), "interrupted")?;
-
-	context.state_store.clear_lease(lease.issue_id())?;
-
-	Ok(true)
-}
-
-fn reconcile_stale_project_lease<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	lease: &IssueLease,
-	issues_by_id: &HashMap<String, TrackerIssue>,
-	cleared_terminal_lane_issue_ids: &mut HashSet<String>,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	let reconciled_status = match issues_by_id.get(lease.issue_id()) {
-		Some(issue) if is_terminal_issue(issue, context.workflow) => "terminated",
-		Some(_) | None => "interrupted",
-	};
-
-	if let Some(issue) = issues_by_id.get(lease.issue_id())
-		&& is_terminal_issue(issue, context.workflow)
-	{
-		clear_terminal_lane_labels_once(
-			context.tracker,
-			context.project,
-			issue,
-			cleared_terminal_lane_issue_ids,
-		)?;
-	}
-
-	mark_run_attempt_if_active(context.state_store, lease.run_id(), reconciled_status)?;
-
-	context.state_store.clear_lease(lease.issue_id())
-}
-
-fn cleanup_missing_orphaned_project_worktree_mappings<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	leases: &[IssueLease],
-	worktrees: &[WorktreeMapping],
-	issues_by_id: &HashMap<String, TrackerIssue>,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	let leased_issue_ids = leases.iter().map(IssueLease::issue_id).collect::<HashSet<_>>();
-
-	for mapping in worktrees {
-		if leased_issue_ids.contains(mapping.issue_id())
-			|| mapping.provenance().is_legacy_unknown()
-			|| !worktree_mapping_path_is_missing(mapping.worktree_path())
-		{
-			continue;
-		}
-
-		let Some(issue) = issues_by_id.get(mapping.issue_id()) else {
-			continue;
-		};
-
-		if issue_has_service_ownership(context.tracker, issue, context.project.service_id())?
-			|| issue.has_label(context.workflow.frontmatter().tracker().needs_attention_label())
-			|| context
-				.state_store
-				.issue_has_active_shared_claim(context.project.service_id(), &issue.id)?
-			|| issue_has_running_attempt(context.state_store, &issue.id)?
-			|| context
-				.state_store
-				.review_handoff_marker(
-					context.project.service_id(),
-					mapping.issue_id(),
-					mapping.branch_name(),
-				)?
-				.is_some()
-		{
-			continue;
-		}
-
-		context.state_store.clear_worktree(mapping.issue_id())?;
-	}
-
-	Ok(())
-}
-
-fn worktree_mapping_path_is_missing(worktree_path: &Path) -> bool {
-	matches!(worktree_path.try_exists(), Ok(false))
-}
-
-fn issue_has_running_attempt(state_store: &StateStore, issue_id: &str) -> Result<bool> {
-	Ok(state_store
-		.latest_run_attempt_for_issue(issue_id)?
-		.is_some_and(|attempt| matches!(attempt.status(), "starting" | "running")))
-}
-
-fn reconcile_orphaned_active_worktree_runs<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	leases: &[IssueLease],
-	worktrees: &[WorktreeMapping],
-	issues_by_id: &HashMap<String, TrackerIssue>,
-	now_unix_epoch: i64,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	let mut orphaned_actions = Vec::new();
-
-	for mapping in worktrees {
-		if leases.iter().any(|lease| lease.issue_id() == mapping.issue_id()) {
-			continue;
-		}
-
-		let Some(issue) = issues_by_id.get(mapping.issue_id()) else {
-			continue;
-		};
-		let Some(action) = inspect_orphaned_active_worktree_reconciliation(
-			context,
-			issue,
-			mapping,
-			now_unix_epoch,
-		)?
-		else {
-			continue;
-		};
-
-		orphaned_actions.push(action);
-	}
-
-	apply_run_lease_reconciliation(
-		context.tracker,
-		context.project,
-		context.state_store,
-		context.worktree_manager,
-		orphaned_actions,
-	)
-}
-
-fn cleanup_terminal_project_worktrees<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	worktrees: &[WorktreeMapping],
-	issues_by_id: &HashMap<String, TrackerIssue>,
-	cleared_terminal_lane_issue_ids: &mut HashSet<String>,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	for mapping in worktrees {
-		if let Some(issue) = issues_by_id.get(mapping.issue_id())
-			&& is_terminal_issue(issue, context.workflow)
-			&& !terminal_issue_keeps_retained_closeout(
-				context.tracker,
-				issue,
-				context.project,
-				context.workflow,
-				context.state_store,
-			)? {
-			clear_terminal_lane_labels_once(
-				context.tracker,
-				context.project,
-				issue,
-				cleared_terminal_lane_issue_ids,
-			)?;
-			cleanup_worktree_mapping(
-				context.state_store,
-				context.worktree_manager,
-				context.workflow,
-				&issue.identifier,
-				mapping,
-			)?;
-		}
-	}
-
-	Ok(())
-}
-
-fn inspect_orphaned_active_worktree_reconciliation<T>(
-	context: &ProjectStateReconciliationContext<'_, T>,
-	issue: &TrackerIssue,
-	worktree_mapping: &WorktreeMapping,
-	now_unix_epoch: i64,
-) -> Result<Option<RunLeaseReconciliation>>
-where
-	T: IssueTracker,
-{
-	let has_service_ownership =
-		issue_has_service_ownership(context.tracker, issue, context.project.service_id())?;
-	let needs_attention =
-		issue.has_label(context.workflow.frontmatter().tracker().needs_attention_label());
-
-	if !has_service_ownership && !needs_attention {
-		return Ok(None);
-	}
-
-	let Some(run_attempt) = context.state_store.latest_run_attempt_for_issue(&issue.id)? else {
-		return Ok(None);
-	};
-	let Some(idle_for) = orphaned_run_lease_idle_duration(
-		context.state_store,
-		&run_attempt,
-		worktree_mapping,
-		now_unix_epoch,
-	)?
-	else {
-		return Ok(None);
-	};
-	let disposition = if needs_attention {
-		RunLeaseDisposition::StalledAlreadyNeedsAttention { idle_for }
-	} else if is_issue_in_progress_for_run(issue, context.workflow)
-		&& worktree_has_tracked_changes(worktree_mapping.worktree_path())
-	{
-		RunLeaseDisposition::StalledRetainedPartialProgress { idle_for }
-	} else if is_issue_in_progress_for_run(issue, context.workflow) {
-		RunLeaseDisposition::Stalled { idle_for }
-	} else {
-		return Ok(None);
-	};
-
-	Ok(Some(RunLeaseReconciliation {
-		issue: issue.clone(),
-		run_attempt,
-		worktree_mapping: Some(worktree_mapping.clone()),
-		disposition,
-		workflow: context.workflow.clone(),
-	}))
-}
-
-fn orphaned_run_lease_idle_duration(
-	state_store: &StateStore,
-	run_attempt: &RunAttempt,
-	worktree_mapping: &WorktreeMapping,
-	now_unix_epoch: i64,
-) -> Result<Option<Duration>> {
-	if !matches!(run_attempt.status(), "starting" | "running") {
-		return Ok(None);
-	}
-
-	let marker = state::read_run_activity_marker_snapshot(worktree_mapping.worktree_path())?
-		.filter(|marker| {
-			marker.run_id() == run_attempt.run_id()
-				&& marker.attempt_number() == run_attempt.attempt_number()
-		});
-
-	if let Some(marker) = marker.as_ref()
-		&& marker.process_id().is_some()
-	{
-		if marker_process_is_alive(marker) {
-			return Ok(None);
-		}
-
-		return Ok(Some(
-			marker
-				.last_activity_unix_epoch()
-				.and_then(|last_activity| observed_idle_duration(last_activity, now_unix_epoch))
-				.unwrap_or(Duration::ZERO),
-		));
-	}
-
-	stalled_idle_duration(state_store, run_attempt, Some(worktree_mapping), now_unix_epoch)
-}
-
-fn clear_terminal_lane_labels_once<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	issue: &TrackerIssue,
-	cleared_issue_ids: &mut HashSet<String>,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	if cleared_issue_ids.insert(issue.id.clone()) {
-		tracker::clear_automation_lane_labels(tracker, issue, project.service_id())?;
-	}
-
-	Ok(())
-}
-
-fn retained_review_lease_matches_run(state_store: &StateStore, lease: &IssueLease) -> Result<bool> {
-	let Some(run_attempt) = state_store.run_attempt(lease.run_id())? else {
-		return Ok(false);
-	};
-	let worktree_mapping = state_store.worktree_for_issue(lease.issue_id())?;
-
-	retained_review_handoff_matches_run(state_store, &run_attempt, worktree_mapping.as_ref())
-}
-
-fn terminal_issue_keeps_retained_closeout<T>(
-	tracker: &T,
-	issue: &TrackerIssue,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-) -> Result<bool>
-where
-	T: IssueTracker + ?Sized,
-{
-	if !is_terminal_issue(issue, workflow) {
-		return Ok(false);
-	}
-
-	Ok(issue_passes_closeout_dispatch_policy(tracker, issue, project, workflow, state_store)?
-		|| closeout_dispatch_block_reason(tracker, issue, project, workflow, state_store)?
-			.is_some())
-}
-
-fn retained_closeout_lease_has_fresh_activity(
-	lease: &IssueLease,
-	issue: &TrackerIssue,
-	project: &ServiceConfig,
-	now_unix_epoch: i64,
-) -> Result<bool> {
-	let worktree_manager =
-		WorktreeManager::new(project.service_id(), project.repo_root(), project.worktree_root());
-	let worktree = worktree_manager.plan_for_issue(&issue.identifier);
-	let Some(marker) = state::read_run_activity_marker_snapshot(&worktree.path)? else {
-		return Ok(false);
-	};
-
-	Ok(marker.run_id() == lease.run_id()
-		&& worktree_activity_marker_is_fresh(&marker, now_unix_epoch))
 }
